requirements.txt
streamlit
pdfplumber
transformers
torch
sentence-transformers
scikit-learn
openai
python-dotenv
openai==1.90.0
